+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add new field $height for EnvelopeTag and EnvelopeTemplateTag
 - Add `EnvelopeTag.recipientClaim`, `EnvelopeTemplateTag.recipientClaim`
 - Add new fields $identificationNumber, $address for `EnvelopeRecipient` and `EnvelopeTemplateRecipient`
+- Add signatureImageContent endpoint to AccountMe 
+- Add new field $hasSignatureImage to User
 
 ## [1.11.0]
 ### Added